Milestone #2 Write-Up
Ok, I make it so it is easier to read. Anyway, I completed the first two sections of the write up, the description and prototype. This is just a first draft that I write, so there are probably grammar errors and wrong details since the prototype isn't complete yet, so you guys can change the write-up anytime and add last three sections once we have people test our programs.