Start of the Mouse Code
This is the start of the mouse code, mostly compliant with the 30%
prototype guidelines. As of now, clicking will create a new note with
the right beat and pitch depending on what the x and y are in the mouse.
I also made it easy to move and stretch the staff, so that when we move
the grid to the middle of the program, it won't mess anything up.