fixed button issues and added data definitions
I fix the code so clicking anywhere on the play button can play the sound and clicking anywhere on reset button can reset the grid. I add data definition for most if not all the functions. Also, I just delete some currently unnecessary codes such as the pausebar and pause-button function for right now to make it looks simpler.